Moved controllers and models into their own files
