+++ controllers/cashflow_controller.py
@@ -1,0 +1,14 @@
+from models.cashflow_model import CashflowModel
+
+class CashflowController:
+    @staticmethod
+    def cashflows():
+        return CashflowModel.get_cashflows()
+    
+    @staticmethod
+    def add_cashflow(expenseid, incomeid, type_):
+        CashflowModel.add_cashflow(expenseid, incomeid, type_)
+
+    @staticmethod
+    def get_cashflow_types():
+        return ['Business', 'Transfer']
