+++ models/budget_model.py
@@ -1,0 +1,46 @@
+from db import db_commit, get_db_connection
+
+class BudgetModel:
+    @staticmethod
+    def get_budgets(year, month):
+        with get_db_connection() as conn:
+            cursor = conn.cursor(dictionary=True)
+            cursor.execute("""
+                           SELECT b.*, c.categoryname, c.type_
+                           FROM budget b
+                           JOIN category c ON b.categoryid = c.categoryid
+                           WHERE b.budget_year = %s AND b.budget_month = %s
+                           ORDER BY c.categoryname
+                           """, (year, month))
+            budgets = cursor.fetchall()
+            # Calculate actual spending for each budget
+            for budget in budgets:
+                cursor.execute("""
+                               SELECT COALESCE(SUM(amount), 0) as actual
+                               FROM transact t
+                               WHERE t.categoryid = %s 
+                               AND YEAR(t.transactiondate) = %s 
+                               AND MONTH(t.transactiondate) = %s
+                               """, (budget['categoryid'], year, month))
+                
+                # execute() returns dict with only key 'actual'
+                actual = cursor.fetchone()['actual']
+                budget['actual'] = actual
+
+            return budgets
+        
+    
+    @staticmethod
+    def add_budget(category_id, budget_year, budget_month, budget_amount):
+        db_commit(
+            """
+                INSERT INTO budget (categoryid, budget_year, 
+                budget_month, budget_amount) 
+                VALUES (%s, %s, %s, %s)
+                ON DUPLICATE KEY UPDATE budget_amount = %s
+            """,
+            (
+                category_id, budget_year, budget_month, budget_amount, 
+                budget_amount
+            )
+        )
